refactor: AI 통화 관련 상수들을 constants 디렉토리로 분리
- VIBRATE_LIST, CALLBACK_LIST, VOICE_LIST를 constants/aiCall.ts로 이동
- 관련 컴포넌트와 유틸 파일에서 해당 상수를 import하도록 수정
