+++ frontend/src/constants/aiCall.ts
@@ -1,0 +1,19 @@
+export const VIBRATE_LIST = [
+  {label: 'Basic', pattern: [1, 500, 200, 500]},
+  {label: 'HeartBeat', pattern: [1, 100, 100, 300, 100, 700]},
+];
+
+export const CALLBACK_LIST = [
+  {label: '10분 후', value: 10},
+  {label: '20분 후', value: 20},
+  {label: '30분 후', value: 30},
+  {label: '40분 후', value: 40},
+  {label: '50분 후', value: 50},
+  {label: '1시간 후', value: 60},
+];
+
+export const VOICE_LIST = [
+  {label: 'Sol', description: '차분함', id: 'sol'},
+  {label: 'Neo', description: '섬세함', id: 'neo'},
+  {label: 'Gina', description: '유쾌함', id: 'gina'},
+];
