+++ frontend/src/store/aiCallSettingsStore.ts
@@ -2,30 +2,8 @@ import AsyncStorage from '@react-native-async-storage/async-storage';
 import {create} from 'zustand';
 import {createJSONStorage, persist} from 'zustand/middleware';
 
+import {CALLBACK_LIST, VIBRATE_LIST, VOICE_LIST} from '../constants/aiCall';
 import {updateScheduledAlarms} from '../utils/alarmManager';
-
-// Vibrate 옵션: label + 실제 패턴
-export const VIBRATE_LIST = [
-  {label: 'Basic', pattern: [1, 500, 200, 500]},
-  {label: 'HeartBeat', pattern: [1, 100, 100, 300, 100, 700]},
-];
-
-// CallBack 옵션: label + 분 단위 value
-export const CALLBACK_LIST = [
-  {label: '10분 후', value: 10},
-  {label: '20분 후', value: 20},
-  {label: '30분 후', value: 30},
-  {label: '40분 후', value: 40},
-  {label: '50분 후', value: 50},
-  {label: '1시간 후', value: 60},
-];
-
-// Voice 옵션: label + description + id
-export const VOICE_LIST = [
-  {label: 'Sol', description: '차분함', id: 'sol'},
-  {label: 'Neo', description: '섬세함', id: 'neo'},
-  {label: 'Gina', description: '유쾌함', id: 'gina'},
-];
 
 export interface AiCallSettingsState {
   selectedDays: number[]; // 0:월 ~ 6:일
